P5-3+P5-6: Decompose card-detail + notes-view god components
card-detail.tsx: 550→232 LOC (−58%) — extracted CardDetailSidebar
(sidebar metadata: priority, column, assignee, blockers, due date, actions)

notes-view.tsx: 1033→532 LOC (−49%) — extracted 5 sub-components:
- buildFolderTree.ts (46 LOC) — pure folder tree builder + FolderNode type
- NoteListItem.tsx (75 LOC) — memo'd note row with dropdown menu
- ArchivedNoteListItem.tsx (41 LOC) — archived note row
- FolderTreeNode.tsx (130 LOC) — recursive folder tree node
- FolderPickerDialog.tsx (221 LOC) — folder picker dialog + recursive node

Also fixed text-[10.5px] → text-[0.75rem] in archived toggle, and
slash-alpha border syntax in blocker chips.

type-check:all, lint, 430 tests pass.
