+++ src/components/notes/notes-view.tsx
@@ -1,15 +1,14 @@
 "use client";
 
-import React, { useState, useEffect, useRef, useCallback, memo, useMemo } from "react";
+import React, { useState, useEffect, useRef, useCallback, useMemo } from "react";
 import {
-  FileText, Plus, Pin, PinOff, Trash2, MoreHorizontal, Search,
-  Wand2, Archive, ArchiveRestore, FolderInput,
-  ChevronDown, ChevronRight, LayoutDashboard, Folder, FolderOpen,
-  FolderPlus, FolderSymlink,
+  FileText, Plus, Search, Wand2,
+  LayoutDashboard, FolderPlus,
+  ChevronDown, ChevronRight, Trash2,
 } from "lucide-react";
 import { useCairnStore } from "@/store";
 import { useShallow } from "zustand/react/shallow";
-import { cn, formatRelative } from "@/lib/utils";
+import { cn } from "@/lib/utils";
 import { Button } from "@/components/ui/button";
 import { Tooltip } from "@/components/ui/tooltip";
 import { NoteEditor } from "./note-editor";
@@ -18,64 +17,15 @@ import { PrdModal } from "./notes-view/PrdModal";
 import { MoveNoteModal } from "./notes-view/MoveNoteModal";
 import { DashboardTemplateModal } from "./DashboardTemplateModal";
 import { useNoteFilter } from "./notes-view/useNoteFilter";
+import { buildFolderTree, type FolderNode } from "./notes-view/buildFolderTree";
+import { NoteListItem } from "./notes-view/NoteListItem";
+import { ArchivedNoteListItem } from "./notes-view/ArchivedNoteListItem";
+import { FolderTreeNode } from "./notes-view/FolderTreeNode";
+import { FolderPickerDialog } from "./notes-view/FolderPickerDialog";
 import type { Note } from "@/types";
-import {
-  DropdownMenu, DropdownMenuTrigger, DropdownMenuContent,
-  DropdownMenuItem, DropdownMenuSeparator,
-} from "@/components/ui/dropdown";
 import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogClose } from "@/components/ui/dialog";
 
-// ── Folder tree helpers ───────────────────────────────────────────────────────
-
-interface FolderNode {
-  name: string;        // segment name (last part of path)
-  path: string;        // full path e.g. "Design/Typography"
-  notes: Note[];
-  children: FolderNode[];
-}
-
-/** Build a tree from a flat list of notes using their folder field. */
-function buildFolderTree(notes: Note[]): { rootNotes: Note[]; folders: FolderNode[] } {
-  const rootNotes: Note[] = [];
-  const folderMap = new Map<string, FolderNode>();
-
-  // Pass 1 — create every folder node (including all ancestors) and
-  // assign notes to their leaf folder. Do NOT wire parent→child here
-  // because the parent node may not exist yet when we encounter a deep path.
-  for (const note of notes) {
-    const folder = note.folder ?? "";
-    if (!folder) {
-      rootNotes.push(note);
-      continue;
-    }
-    const segments = folder.split("/").filter(Boolean);
-    let built = "";
-    for (const seg of segments) {
-      built = built ? `${built}/${seg}` : seg;
-      if (!folderMap.has(built)) {
-        folderMap.set(built, { name: seg, path: built, notes: [], children: [] });
-      }
-    }
-    folderMap.get(folder)!.notes.push(note);
-  }
-
-  // Pass 2 — wire every node into its parent now that all nodes exist.
-  for (const node of folderMap.values()) {
-    const lastSlash = node.path.lastIndexOf("/");
-    if (lastSlash === -1) continue; // top-level — no parent to wire
-    const parentPath = node.path.slice(0, lastSlash);
-    folderMap.get(parentPath)?.children.push(node);
-  }
-
-  // Collect top-level folders and sort them alphabetically.
-  const topLevel: FolderNode[] = [];
-  for (const node of folderMap.values()) {
-    if (!node.path.includes("/")) topLevel.push(node);
-  }
-  topLevel.sort((a, b) => a.name.localeCompare(b.name));
-
-  return { rootNotes, folders: topLevel };
-}
+// ── NotesView orchestrator ──────────────────────────────────────────────────
 
 export function NotesView() {
   const {
@@ -461,7 +411,7 @@ export function NotesView() {
           {archivedNotes.length > 0 && (
             <div className="mt-1 border-t border-[var(--border-subtle)]">
               <button onClick={() => setShowArchivedNotes((v) => !v)}
-                className="flex items-center gap-1.5 w-full px-3 py-2 text-[10.5px] text-[var(--text-tertiary)] hover:text-[var(--text-secondary)] transition-colors">
+                className="flex items-center gap-1.5 w-full px-3 py-2 text-[0.75rem] text-[var(--text-tertiary)] hover:text-[var(--text-secondary)] transition-colors">
                 {showArchivedNotes ? <ChevronDown size={10} /> : <ChevronRight size={10} />}
                 {archivedNotes.length} archived
               </button>
@@ -575,459 +525,7 @@ export function NotesView() {
           </div>
         </DialogContent>
       </Dialog>
-    </div>
-  );
-}
+     </div>
+   );
+ }
 
-// ── FolderTreeNode ────────────────────────────────────────────────────────────
-
-interface FolderTreeNodeProps {
-  node: FolderNode;
-  activeNoteId: string | null;
-  collapsed: Record<string, boolean>;
-  depth?: number;
-  onToggle: (path: string) => void;
-  onNoteClick: (id: string) => void;
-  onNotePin: (note: Note) => void;
-  onNoteDelete: (note: Note) => void;
-  onNoteArchive: (note: Note) => void;
-  onNoteMove: (note: Note) => void;
-  onNoteMoveToFolder: (note: Note) => void;
-  onNoteReveal: (note: Note) => void;
-  onCreateInFolder: (folder: string) => void;
-  // drag-and-drop
-  dropTarget: string | null;
-  onNoteDragStart: (noteId: string) => void;
-  onNoteDragEnd: () => void;
-  onFolderDragOver: (folderPath: string) => void;
-  onFolderDragLeave: () => void;
-  onFolderDrop: (folderPath: string) => void;
-}
-
-const FolderTreeNode = memo(function FolderTreeNode({
-  node, activeNoteId, collapsed, depth = 0,
-  onToggle, onNoteClick, onNotePin, onNoteDelete,
-  onNoteArchive, onNoteMove, onNoteMoveToFolder, onNoteReveal, onCreateInFolder,
-  dropTarget, onNoteDragStart, onNoteDragEnd, onFolderDragOver, onFolderDragLeave, onFolderDrop,
-}: FolderTreeNodeProps) {
-  const isCollapsed = collapsed[node.path];
-  const indent = depth * 10;
-  const isDragOver = dropTarget === node.path;
-
-  return (
-    <div>
-      {/* Folder row */}
-      <div
-        className={cn(
-          "group flex items-center gap-1 px-2 py-1.5 cursor-pointer transition-colors",
-          isDragOver
-            ? "bg-[color-mix(in_srgb,var(--accent)_15%,transparent)] outline outline-1 outline-[var(--accent)] outline-offset-[-1px] rounded"
-            : "hover:bg-[var(--surface-2)]",
-        )}
-        style={{ paddingLeft: `${8 + indent}px` }}
-        onClick={() => onToggle(node.path)}
-        onDragOver={(e) => { e.preventDefault(); e.stopPropagation(); onFolderDragOver(node.path); }}
-        onDragLeave={(e) => { e.stopPropagation(); onFolderDragLeave(); }}
-        onDrop={(e) => { e.preventDefault(); e.stopPropagation(); onFolderDrop(node.path); }}
-      >
-        {isCollapsed
-          ? <ChevronRight size={11} className="text-[var(--text-tertiary)] flex-shrink-0" />
-          : <ChevronDown size={11} className="text-[var(--text-tertiary)] flex-shrink-0" />}
-        {isCollapsed
-          ? <Folder size={12} className="text-[var(--text-tertiary)] flex-shrink-0" />
-          : <FolderOpen size={12} className="text-[var(--accent)] flex-shrink-0" />}
-        <span className="text-[0.786rem] font-medium text-[var(--text-secondary)] flex-1 truncate">{node.name}</span>
-        <button
-          onClick={(e) => { e.stopPropagation(); onCreateInFolder(node.path); }}
-          className="opacity-0 group-hover:opacity-100 p-0.5 rounded text-[var(--text-tertiary)] hover:text-[var(--accent)] hover:bg-[var(--accent-dim)] transition-all"
-          title={`New note in ${node.name}`}
-        >
-          <Plus size={10} />
-        </button>
-      </div>
-
-      {/* Folder contents */}
-      {!isCollapsed && (
-        <div>
-          {/* Child folders */}
-          {node.children.map((child) => (
-            <FolderTreeNode
-              key={child.path}
-              node={child}
-              activeNoteId={activeNoteId}
-              collapsed={collapsed}
-              depth={depth + 1}
-              onToggle={onToggle}
-              onNoteClick={onNoteClick}
-              onNotePin={onNotePin}
-              onNoteDelete={onNoteDelete}
-              onNoteArchive={onNoteArchive}
-              onNoteMove={onNoteMove}
-              onNoteMoveToFolder={onNoteMoveToFolder}
-              onNoteReveal={onNoteReveal}
-              onCreateInFolder={onCreateInFolder}
-              dropTarget={dropTarget}
-              onNoteDragStart={onNoteDragStart}
-              onNoteDragEnd={onNoteDragEnd}
-              onFolderDragOver={onFolderDragOver}
-              onFolderDragLeave={onFolderDragLeave}
-              onFolderDrop={onFolderDrop}
-            />
-          ))}
-          {/* Notes in this folder */}
-          {node.notes.map((note) => (
-            <NoteListItem
-              key={note.id}
-              note={note}
-              isActive={note.id === activeNoteId}
-              indent={indent + 16}
-              onClick={() => onNoteClick(note.id)}
-              onPin={() => onNotePin(note)}
-              onDelete={() => onNoteDelete(note)}
-              onArchive={() => onNoteArchive(note)}
-              onMove={() => onNoteMove(note)}
-              onMoveToFolder={() => onNoteMoveToFolder(note)}
-              onReveal={() => onNoteReveal(note)}
-              onDragStart={onNoteDragStart}
-              onDragEnd={onNoteDragEnd}
-            />
-          ))}
-          {/* Empty folder state */}
-          {node.notes.length === 0 && node.children.length === 0 && (
-            <button
-              onClick={() => onCreateInFolder(node.path)}
-              className="w-full flex items-center gap-1.5 px-2 py-1.5 text-[0.714rem] text-[var(--text-tertiary)] hover:text-[var(--accent)] hover:bg-[var(--surface-2)] transition-colors"
-              style={{ paddingLeft: `${indent + 24}px` }}
-            >
-              <Plus size={10} />
-              New note
-            </button>
-          )}
-        </div>
-      )}
-    </div>
-  );
-});
-
-// ── NoteListItem ──────────────────────────────────────────────────────────────
-
-interface NoteListItemProps {
-  note: Note; isActive: boolean; indent?: number;
-  onClick: () => void; onPin: () => void; onDelete: () => void;
-  onArchive: () => void; onMove: () => void; onMoveToFolder: () => void; onReveal: () => void;
-  onDragStart?: (noteId: string) => void;
-  onDragEnd?: () => void;
-}
-
-const NoteListItem = React.memo(function NoteListItem({ note, isActive, indent = 0, onClick, onPin, onDelete, onArchive, onMove, onMoveToFolder, onReveal, onDragStart, onDragEnd }: NoteListItemProps) {
-  return (
-    <div onClick={onClick}
-      draggable={!!onDragStart}
-      onDragStart={(e) => { e.stopPropagation(); onDragStart?.(note.id); }}
-      onDragEnd={(e) => { e.stopPropagation(); onDragEnd?.(); }}
-      className={cn("group relative flex flex-col gap-0.5 py-2.5 cursor-pointer transition-colors pr-3",
-        isActive ? "bg-[var(--surface-2)] border-l-2 border-[var(--accent)]" : "hover:bg-[var(--surface-2)] border-l-2 border-transparent")}
-      style={{ paddingLeft: `${12 + indent}px` }}
-    >
-      <div className="flex items-center gap-1.5">
-        {note.isPinned && <Pin size={9} className="text-[var(--accent)] flex-shrink-0" />}
-        {note.type === "dashboard" && <LayoutDashboard size={9} className="text-[var(--text-tertiary)] flex-shrink-0" />}
-        <span className={cn("text-xs font-medium truncate flex-1", isActive ? "text-[var(--text-primary)]" : "text-[var(--text-secondary)]")}>
-          {note.title}
-        </span>
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild>
-            <button onClick={(e) => e.stopPropagation()}
-              className="opacity-0 group-hover:opacity-100 p-0.5 rounded text-[var(--text-tertiary)] hover:text-[var(--text-primary)] hover:bg-[var(--surface-3)] transition-all">
-              <MoreHorizontal size={11} />
-            </button>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent align="end">
-            <DropdownMenuItem onClick={(e) => { e.stopPropagation(); onPin(); }}>
-              {note.isPinned ? <PinOff size={12} /> : <Pin size={12} />}
-              {note.isPinned ? "Unpin" : "Pin"}
-            </DropdownMenuItem>
-            <DropdownMenuItem onClick={(e) => { e.stopPropagation(); onReveal(); }}>
-              <FileText size={12} />Reveal in {window.electron?.platform === "win32" ? "Explorer" : window.electron?.platform === "linux" ? "Files" : "Finder"}
-            </DropdownMenuItem>
-            <DropdownMenuSeparator />
-            <DropdownMenuItem onClick={(e) => { e.stopPropagation(); onArchive(); }}>
-              <Archive size={12} />Archive
-            </DropdownMenuItem>
-            <DropdownMenuItem onClick={(e) => { e.stopPropagation(); onMoveToFolder(); }}>
-              <FolderSymlink size={12} />Move to folder
-            </DropdownMenuItem>
-            <DropdownMenuItem onClick={(e) => { e.stopPropagation(); onMove(); }}>
-              <FolderInput size={12} />Move to project
-            </DropdownMenuItem>
-            <DropdownMenuSeparator />
-            <DropdownMenuItem onClick={(e) => { e.stopPropagation(); onDelete(); }} className="text-[var(--danger)] hover:text-[var(--danger)]">
-              <Trash2 size={12} />Delete
-            </DropdownMenuItem>
-          </DropdownMenuContent>
-        </DropdownMenu>
-      </div>
-      <span className="text-[0.786rem] text-[var(--text-tertiary)] truncate">{note.contentText.slice(0, 60) || "Empty note"}</span>
-      <span className="text-[0.714rem] text-[var(--text-tertiary)]">{formatRelative(note.updatedAt)}</span>
-    </div>
-  );
-});
-
-// ── ArchivedNoteListItem ──────────────────────────────────────────────────────
-
-function ArchivedNoteListItem({ note, onRestore, onDelete }: { note: Note; onRestore: () => void; onDelete: () => void }) {
-  return (
-    <div className="group flex items-center gap-1.5 px-3 py-2">
-      <span className="flex-1 text-[0.786rem] text-[var(--text-tertiary)] truncate">{note.title}</span>
-      <Tooltip content="Restore note">
-        <button onClick={(e) => { e.stopPropagation(); onRestore(); }}
-          className="opacity-0 group-hover:opacity-100 p-0.5 rounded text-[var(--text-tertiary)] hover:text-[var(--accent)] hover:bg-[var(--accent-dim)] transition-all">
-          <ArchiveRestore size={11} />
-        </button>
-      </Tooltip>
-      <DropdownMenu>
-        <DropdownMenuTrigger asChild>
-          <button onClick={(e) => e.stopPropagation()}
-            className="opacity-0 group-hover:opacity-100 p-0.5 rounded text-[var(--text-tertiary)] hover:text-[var(--text-primary)] hover:bg-[var(--surface-3)] transition-all">
-            <MoreHorizontal size={11} />
-          </button>
-        </DropdownMenuTrigger>
-        <DropdownMenuContent align="end">
-          <DropdownMenuItem onClick={(e) => { e.stopPropagation(); onRestore(); }}>
-            <ArchiveRestore size={12} />Restore
-          </DropdownMenuItem>
-          <DropdownMenuSeparator />
-          <DropdownMenuItem onClick={(e) => { e.stopPropagation(); onDelete(); }} className="text-[var(--danger)] hover:text-[var(--danger)]">
-            <Trash2 size={12} />Delete permanently
-          </DropdownMenuItem>
-        </DropdownMenuContent>
-      </DropdownMenu>
-    </div>
-  );
-}
-
-// ── FolderPickerDialog ────────────────────────────────────────────────────────
-
-interface FolderPickerDialogProps {
-  folderTree: FolderNode[];
-  /** "move" — clicking a row moves immediately. "create" — picking a row sets parent, inline input creates child. */
-  mode: "move" | "create";
-  currentFolder?: string;
-  onSelect: (folder: string) => void;
-  onClose: () => void;
-}
-
-function FolderPickerDialog({ folderTree, mode, currentFolder = "", onSelect, onClose }: FolderPickerDialogProps) {
-  // parentPath="" means creating at root; "Design" means child of Design
-  const [addingUnder, setAddingUnder] = useState<string | null>(null);
-  const [newFolderName, setNewFolderName] = useState("");
-  const newInputRef = useRef<HTMLInputElement>(null);
-
-  useEffect(() => {
-    if (addingUnder !== null) newInputRef.current?.focus();
-  }, [addingUnder]);
-
-  function handleCreate() {
-    const name = newFolderName.trim();
-    if (!name) return;
-    const full = addingUnder ? `${addingUnder}/${name}` : name;
-    onSelect(full);
-  }
-
-  function openAdd(parentPath: string, e: React.MouseEvent) {
-    e.stopPropagation();
-    setAddingUnder(parentPath);
-    setNewFolderName("");
-  }
-
-  function cancelAdd() {
-    setAddingUnder(null);
-    setNewFolderName("");
-  }
-
-  const inlineInput = (
-    <div className="flex items-center gap-2 px-3 py-2 border-t border-[var(--border)] bg-[var(--surface-2)]">
-      <FolderPlus size={12} className="text-[var(--accent)] shrink-0" />
-      <input
-        ref={newInputRef}
-        value={newFolderName}
-        onChange={(e) => setNewFolderName(e.target.value)}
-        onKeyDown={(e) => {
-          if (e.key === "Enter") handleCreate();
-          if (e.key === "Escape") cancelAdd();
-        }}
-        placeholder="Folder name…"
-        className="flex-1 bg-transparent text-xs text-[var(--text-primary)] placeholder:text-[var(--text-tertiary)] outline-none"
-      />
-      <button
-        onClick={handleCreate}
-        disabled={!newFolderName.trim()}
-        className="text-[0.714rem] text-[var(--accent)] disabled:opacity-30 hover:underline shrink-0"
-      >
-        Create & move
-      </button>
-      <button onClick={cancelAdd} className="text-[0.714rem] text-[var(--text-tertiary)] hover:text-[var(--text-secondary)]">
-        ✕
-      </button>
-    </div>
-  );
-
-  return (
-    <Dialog open onOpenChange={(o) => { if (!o) onClose(); }}>
-      <DialogContent size="sm">
-        <DialogHeader>
-          <DialogTitle>{mode === "create" ? "New folder" : "Move to folder"}</DialogTitle>
-        </DialogHeader>
-        <div className="px-5 pb-5 pt-1">
-          <div className="rounded-lg border border-[var(--border)] overflow-hidden mb-3">
-
-            {/* Root row */}
-            <div className={cn(
-              "group flex items-center gap-2 px-3 py-2 text-xs transition-colors border-b border-[var(--border)]",
-              currentFolder === "" && mode === "move" ? "bg-[var(--accent-dim)] text-[var(--accent)]" : "text-[var(--text-secondary)] hover:bg-[var(--surface-2)]"
-            )}>
-              <button
-                className="flex items-center gap-2 flex-1 min-w-0 text-left"
-                onClick={() => mode === "move" ? onSelect("") : openAdd("", { stopPropagation: () => {} } as React.MouseEvent)}
-              >
-                <Folder size={12} className="shrink-0" />
-                <span className="font-medium">Root</span>
-              </button>
-              <button
-                onClick={(e) => openAdd("", e)}
-                className="opacity-0 group-hover:opacity-100 p-0.5 rounded text-[var(--text-tertiary)] hover:text-[var(--accent)] transition-all"
-                title="New folder in root"
-              >
-                <FolderPlus size={11} />
-              </button>
-            </div>
-
-            {/* Inline input under root */}
-            {addingUnder === "" && inlineInput}
-
-            {folderTree.length === 0 && addingUnder === null && (
-              <p className="px-3 py-3 text-[0.714rem] text-[var(--text-tertiary)]">No folders yet</p>
-            )}
-
-            {/* Recursive tree */}
-            {folderTree.map((node) => (
-              <FolderPickerNode
-                key={node.path}
-                node={node}
-                currentFolder={currentFolder}
-                depth={0}
-                mode={mode}
-                addingUnder={addingUnder}
-                newFolderName={newFolderName}
-                newInputRef={newInputRef}
-                onSelect={onSelect}
-                onOpenAdd={openAdd}
-                onNameChange={setNewFolderName}
-                onCreate={handleCreate}
-                onCancel={cancelAdd}
-                inlineInput={inlineInput}
-              />
-            ))}
-          </div>
-
-          <div className="flex justify-end">
-            <DialogClose asChild>
-              <Button variant="ghost" size="sm">Cancel</Button>
-            </DialogClose>
-          </div>
-        </div>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-interface FolderPickerNodeProps {
-  node: FolderNode;
-  currentFolder: string;
-  depth: number;
-  mode: "move" | "create";
-  addingUnder: string | null;
-  newFolderName: string;
-  newInputRef: React.RefObject<HTMLInputElement | null>;
-  onSelect: (folder: string) => void;
-  onOpenAdd: (parentPath: string, e: React.MouseEvent) => void;
-  onNameChange: (v: string) => void;
-  onCreate: () => void;
-  onCancel: () => void;
-  inlineInput: React.ReactNode;
-}
-
-function FolderPickerNode({
-  node, currentFolder, depth, mode, addingUnder, onSelect, onOpenAdd, inlineInput,
-}: FolderPickerNodeProps) {
-  const [open, setOpen] = useState(true);
-  const isSelected = mode === "move" && currentFolder === node.path;
-  const indent = depth * 12;
-
-  return (
-    <div className="border-t border-[var(--border)]">
-      <div
-        className={cn(
-          "group flex items-center gap-2 text-xs transition-colors",
-          isSelected
-            ? "bg-[var(--accent-dim)] text-[var(--accent)]"
-            : "text-[var(--text-secondary)] hover:bg-[var(--surface-2)]"
-        )}
-        style={{ paddingLeft: `${12 + indent}px`, paddingRight: "12px", paddingTop: "7px", paddingBottom: "7px" }}
-      >
-        {/* Chevron */}
-        {node.children.length > 0 ? (
-          <button
-            onClick={(e) => { e.stopPropagation(); setOpen((o) => !o); }}
-            className="shrink-0 text-[var(--text-tertiary)]"
-          >
-            {open ? <ChevronDown size={10} /> : <ChevronRight size={10} />}
-          </button>
-        ) : (
-          <span className="w-[10px] shrink-0" />
-        )}
-        {/* Folder name */}
-        <button
-          className="flex items-center gap-2 flex-1 min-w-0 text-left"
-          onClick={(e) => mode === "move" ? onSelect(node.path) : onOpenAdd(node.path, e)}
-        >
-          {open && node.children.length > 0
-            ? <FolderOpen size={12} className="shrink-0" />
-            : <Folder size={12} className="shrink-0" />}
-          <span className="truncate">{node.name}</span>
-        </button>
-        {/* Add child folder button */}
-        <button
-          onClick={(e) => onOpenAdd(node.path, e)}
-          className="opacity-0 group-hover:opacity-100 p-0.5 rounded text-[var(--text-tertiary)] hover:text-[var(--accent)] transition-all shrink-0"
-          title={`New folder in ${node.name}`}
-        >
-          <FolderPlus size={11} />
-        </button>
-      </div>
-
-      {/* Inline input for this node */}
-      {addingUnder === node.path && inlineInput}
-
-      {/* Children */}
-      {open && node.children.map((child) => (
-        <FolderPickerNode
-          key={child.path}
-          node={child}
-          currentFolder={currentFolder}
-          depth={depth + 1}
-          mode={mode}
-          addingUnder={addingUnder}
-          newFolderName=""
-          newInputRef={{ current: null }}
-          onSelect={onSelect}
-          onOpenAdd={onOpenAdd}
-          onNameChange={() => {}}
-          onCreate={() => {}}
-          onCancel={() => {}}
-          inlineInput={inlineInput}
-        />
-      ))}
-    </div>
-  );
-}
